Added some codes.  Removed the task category code.
Added the codes for: IT Security training, Accessibility training and awareness, official language training, and the code for 'improving developer productivity'.  
Removed the task code for general training as this results in training be allocated to 'non-work activities'.  Added a note to say that the continuous improvement code can also be used for general training. Need to talk to the CATS and learning teams about this.