Continuous improvement update (#853)
* Adding recommended cloud certs

* Adding reduction of technical debt to improve and added CATS code for grassroots

* I have no idea how to do sub lists ... travis hates me

* Minor changes

* Add Learning CATS code

* Add CATS for Career development

* Add CATS code for IM Clean-up

* Added some codes.  Removed the task category code.

Added the codes for: IT Security training, Accessibility training and awareness, official language training, and the code for 'improving developer productivity'.  
Removed the task code for general training as this results in training be allocated to 'non-work activities'.  Added a note to say that the continuous improvement code can also be used for general training. Need to talk to the CATS and learning teams about this.

* Added more task component codes back in

* trying to get Travis to like me

Co-authored-by: Gabriel Cossette <gabriel.cossette@hrsdc-rhdcc.gc.ca>
Co-authored-by: RachelMuston <rachel_muston@hotmail.com>